TDC
Added support for The Drowned City player cards (stater deck are of placeholder quality for now - legal but meh).

Fixed the number of Skull chaos tokens in Curse of the Rougarou (Hard mode).

Added support for loading decks where an investigator was transformed into the Lost Homunculus (pending PR integration and deployment).

Added the option to spawn new cards in Player 1-4 Investigator groups as well as in the Encounter Deck group.

Added functionality to modify game variables in the same way as player variables (for Eye of Ghathanothoa).

Added functionality to set default hand limit, which is different to modifying default hand limit by X (for George Barnaby).

Updated attach code to use latest API (set attachment direction when attaching, not in the next step).